Improve Ollama models
They're not as good as ChatGPT to break down nested function calling, hallucinating values for dependent function calls.
So to force them to slow down, only process the 1st function call, and rewrite the assistant message to discard the calls that weren't made.
This way, the model will repeat subsequent nested calls after the values from the call they depend on is available, leading to better arguments.